feat: Got all util functions under volt_utils
