+++ volt_add/Cargo.toml
@@ -14,3 +14,4 @@ indicatif = "0.16"
 tokio = { version = "1.5.0", features = ["full"] }
 volt_core = { path = "../volt_core" }
 volt_init = { path = "../volt_init" }
+volt_utils = {path="../volt_utils"}
