+++ volt_cache/Cargo.toml
@@ -11,3 +11,4 @@ async-trait = "0.1"
 colored = "2.0"
 indicatif = "0.16"
 volt_core = { path = "../volt_core" }
+volt_utils = {path= "../volt_utils"}
