+++ volt_cache/src/command.rs
@@ -23,9 +23,9 @@ use anyhow::Result;
 use async_trait::async_trait;
 use colored::Colorize;
 use indicatif::{ProgressBar, ProgressStyle};
-use volt_core::app::App;
+use volt_utils::app::App;
 use volt_core::command::Command;
-use volt_core::utils::PROGRESS_CHARS;
+use volt_utils::PROGRESS_CHARS;
 use volt_core::VERSION;
 
 /// Struct implementation for the `Add` command.
