+++ volt_clone/Cargo.toml
@@ -10,3 +10,4 @@ anyhow = "1.0"
 async-trait = "0.1"
 colored = "2.0"
 volt_core = { path = "../volt_core" }
+volt_utils = {path = "../volt_utils"}
