+++ volt_compress/src/command.rs
@@ -18,8 +18,8 @@ use std::sync::Arc;
 use anyhow::Result;
 use async_trait::async_trait;
 use colored::Colorize;
-use volt_core::{app::App, command::Command, VERSION};
-
+use volt_core::{command::Command, VERSION};
+use volt_utils::app::App;
 pub struct Compress {}
 
 #[async_trait]
