+++ volt_core/src/command.rs
@@ -19,7 +19,7 @@ use std::sync::Arc;
 use anyhow::Result;
 use async_trait::async_trait;
 
-use crate::app::App;
+use volt_utils::app::App;
 
 #[async_trait]
 pub trait Command {
