+++ volt_core/src/lib.rs
@@ -17,13 +17,11 @@
 #[macro_use]
 extern crate lazy_static;
 
-pub mod app;
 pub mod classes;
 pub mod command;
 pub mod io;
 pub mod model;
 pub mod prompt;
-pub mod utils;
 
 // Constants
 pub const VERSION: &str = env!("CARGO_PKG_VERSION");
