+++ volt_deploy/src/command.rs
@@ -21,8 +21,8 @@ use std::{env, process, sync::Arc};
 use anyhow::Result;
 use async_trait::async_trait;
 use colored::Colorize;
-use volt_core::{app::App, command::Command, VERSION};
-
+use volt_core::{command::Command, VERSION};
+use volt_utils::app::App;
 /// Struct implementation for the `Deploy` command.
 pub struct Deploy;
