+++ volt_migrate/src/command.rs
@@ -22,10 +22,10 @@ use anyhow::Result;
 use async_trait::async_trait;
 use colored::Colorize;
 use volt_core::{
-    app::App, classes::package_manager::PackageManager, command::Command, prompt::prompts::Select,
+   classes::package_manager::PackageManager, command::Command, prompt::prompts::Select,
     VERSION,
 };
-
+use volt_utils::app::App;
 /// Struct implementation for the `Migrate` command.
 pub struct Migrate;
