+++ volt_unknown/src/command.rs
@@ -18,8 +18,9 @@ use std::sync::Arc;
 use anyhow::Result;
 use async_trait::async_trait;
 use colored::Colorize;
-use volt_core::{app::App, classes::package::PackageJson, command::Command};
-
+use volt_core::{command::Command};
+use volt_utils::package::PackageJson;
+use volt_utils::app::App;
 pub struct Unknown {}
 
 #[async_trait]
