+++ volt_utils/Cargo.toml
@@ -1,9 +1,10 @@
 [package]
-name = "volt_core"
-version = "0.0.1"
-authors = ["Volt Contributors (https://github.com/voltpkg/volt/graphs/contributors)"]
-description = "Volt's core library."
+name = "volt_utils"
+version = "0.1.0"
+authors = ["VarunPotti <potti.varun07@gmail.com>"]
 edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 anyhow = "1.0"
@@ -22,7 +23,7 @@ structopt = "0.3"
 tar = "0.4"
 thiserror = "1.0"
 tokio = { version = "1.5.0", features = ["full"] }
-volt_utils = { path = "../volt_utils" }
+indicatif = "0.16"
 
 [target.'cfg(windows)'.dependencies]
 winapi = { version = "0.3", features = ["errhandlingapi", "fileapi", "guiddef", "handleapi", "ioapiset", "processthreadsapi", "securitybaseapi", "winbase", "winioctl", "winnt"]}
